+++ metatensor-torch/src/block.cpp
@@ -36,6 +36,7 @@ TensorBlockHolder::TensorBlockHolder(
         /* parent */ torch::IValue()
     )
 {
+    
     auto values_device = this->values().device();
     if (values_device != this->samples()->values().device()) {
         C10_THROW_ERROR(ValueError,