+++ python/metatensor-core/metatensor/block.py
@@ -182,6 +182,20 @@ class TensorBlock:
             "TensorMap first"
         )
 
+    def __len__(self) -> int:
+        """
+        Get the length of the values stored in this block
+        (i.e. the number of samples in the block)
+        """
+        return len(self.values)
+
+    @property
+    def shape(self):
+        """
+        Get the shape of the values  array in this block.
+        """
+        return self.values.shape
+
     def copy(self) -> "TensorBlock":
         """
         get a deep copy of this block, including all the data and metadata